requirements.txt
python-dotenv>=1.0.0
streamlit>=1.28.0
langchain-classic>=1.0.0
langchain-core>=0.1.0
requests==2.31.0